[ty] Extract implicit-attribute inference into a dedicated module
